Bugfix: Integer truncation on 32 bit systems.

+++ include/osmium/util/memory.hpp
@@ -47,7 +47,7 @@ namespace osmium {
         int64_t m_peak    = 0;
 
 #ifdef __linux__
-        static int parse_number(const std::string& line) {
+        static int64_t parse_number(const std::string& line) {
             const auto f = line.find_first_of("0123456789");
             const auto l = line.find_last_of("0123456789");
             return osmium::detail::str_to_int<int64_t>(line.substr(f, l - f + 1).c_str());
